mlproject/requirements.txt
uvicorn
fastapi
pydantic
mlflow
scikit-learn
pandas
numpy
python-dotenv
psycopg2-binary